@@ -75,7 +75,7 @@ public record Loan : LiabilityBase {
     public ICollection<DateValue> Repayments {
         get;
         init;
-    } = new List<DateValue>();
+    } = [];
 
     /// <summary>
     /// Tilgungsintervall
@@ -91,7 +91,7 @@ public record Loan : LiabilityBase {
     public ICollection<DateValue> AdditionalRepayments {
         get;
         init;
-    } = new List<DateValue>();
+    } = [];
 
     /// <summary>
     /// Zinssätze in %
@@ -99,7 +99,7 @@ public record Loan : LiabilityBase {
     public ICollection<PercentValue> InterestRates {
         get;
         init;
-    } = new List<PercentValue>();
+    } = [];
 
     /// <summary>
     /// Ende der Zinsbindung
@@ -115,7 +115,7 @@ public record Loan : LiabilityBase {
     public ICollection<DateValue> InterestGrants {
         get;
         init;
-    } = new List<DateValue>();
+    } = [];
 
     /// <summary>
     /// Teilauszahlungen der Darlehenssumme
@@ -123,7 +123,7 @@ public record Loan : LiabilityBase {
     public ICollection<DateValue> LoanPayouts {
         get;
         init;
-    } = new List<DateValue>();
+    } = [];
 
     /// <summary>
     /// Auszahlung zu Darlehensbeginn
